+++ Source/Core/Visualization/Viewer/Application.h
@@ -13,13 +13,17 @@
  */
 /*===========================================================================*/
 
-#if defined( KVS_APP_GLUT ) && defined( KVS_SUPPORT_GLUT )
+#if defined( KVS_SUPPORT_GLUT ) && defined( KVS_APP_USE_GLUT )
 #include <kvs/glut/Application>
 namespace kvs { using Application = kvs::glut::Application; }
 
-#elif defined( KVS_APP_GLFW ) && defined( KVS_SUPPORT_GLFW )
+#elif defined( KVS_SUPPORT_GLFW ) && defined( KVS_APP_USE_GLFW )
 #include <kvs/glfw/Application>
 namespace kvs { using Application = kvs::glfw::Application; }
+
+#elif defined( KVS_SUPPORT_QT ) && defined( KVS_APP_USE_QT )
+#include <kvs/qt/Application>
+namespace kvs { using Application = kvs::qt::Application; }
 
 #else
 
@@ -28,6 +32,10 @@ namespace kvs { using Application = kvs::glfw::Application; }
 namespace kvs { using Application = kvs::glfw::Application; }
 
 #elif defined( KVS_SUPPORT_GLUT )
+#include <kvs/glut/Application>
+namespace kvs { using Application = kvs::glut::Application; }
+
+#elif defined( KVS_SUPPORT_QT )
 #include <kvs/glut/Application>
 namespace kvs { using Application = kvs::glut::Application; }